requirements.txt
# ─── Web Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ─── Authentication ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ─── MongoDB Atlas ────────────────────────────────────────────
pymongo[srv]==4.7.3
dnspython==2.6.1

# ─── PDF Processing ───────────────────────────────────────────
PyPDF2==3.0.1
pdfminer.six==20231228

# ─── Vector Database ──────────────────────────────────────────
chromadb==0.5.3

# ─── ML / NLP ─────────────────────────────────────────────────
scikit-learn==1.5.0
numpy==1.26.4
networkx==3.3
scipy==1.13.1

# ─── LLM (Groq — free, fast) ──────────────────────────────────
groq==0.9.0

# ─── Local Embeddings (no API key needed) ────────────────────
sentence-transformers==3.0.1
torch==2.3.1        # CPU-only is fine; sentence-transformers needs torch

# ─── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.3
aiofiles==23.2.1
loguru==0.7.2
